claude: try to dry out the custom gov signer, but mess everything up
